Update README with info on new repository
The Exponent SDK has all the native components and API's that I need to build this app, and as a bonus they are all cross platform compatible with iOS and Android!  I've made a new project repository to port the app over to Exponent. Future work will continue in the new repository: https://github.com/Traviskn/react-native-exponent-pokemon-go